+++ eco_perf/terminal_interface/io/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_library(io
     format_definitions.h format_definitions.c
     io.h io.c
+    string_utils.h string_utils.c
 )
