+++ eco_perf/terminal_interface/io/string_utils.h
@@ -1,0 +1,7 @@
+#ifndef ECO_PERF_STRING_UTILS_H_INCLUDED
+#define ECO_PERF_STRING_UTILS_H_INCLUDED
+
+char *str_append(char *destination, char const *src);
+char *fill_str(char *destination, char filler, int n_repeat);
+
+#endif
